// _backup_original/backup_20260225_1503/_backup_20260221_2250/types.ts
export interface DashboardData {
  consolidatedBalance: number;
  netWorth: number;
  creditCards: {
    brand: string;
    current: number;
    forecasted: number;
    color: string;
  }[];
  alerts: {
    id: string;
    type: 'warning' | 'info' | 'critical';
    message: string;
    createdAt: string;
  }[];
  goals: {
    id: string;
    name: string;
    target: number;
    current: number;
    color: string;
  }[];
  cashFlow: {
    month: string;
    income: number;
    expense: number;
  }[];
  assets: {
    name: string;
    value: number;
    color: string;
  }[];
}